[build] Replace profile_list with profile_source_files and profile_source_files_list_files
This gives a clearer name and documentation to the GN build argument
for providing a prepopulated file for selective instrumentation, and
makes it a list of such files.  It also adds another build argument
that takes the list of source files (as GN paths) directly and
populates a file to feed to the compiler.

Bug: 68518
Change-Id: Idb61d7fdcd1175501c420b9a1cd91a9e181d4833
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/477513
Commit-Queue: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Petr Hosek <phosek@google.com>