[hlcpp] Call error handler in proxy controller in Send()
Currently, zx_status_t is returned but it isn't read anywhere. This
means that Write() and Encode() errors are inaccessible. Because
it is challenging to change the callers of Send() to return a
zx_status_t, this CL re-uses an existing error handler mechanism.

Most error handlers work when client side errors are enabled,
however, there are a subset of cases where this is not the case
and my efforts to create direct fixes were fruitless. As a result,
I created an RAII object
"TransitoryProxyControllerClientSideErrorDisabler"
that disables client side error reporting and added it to relevant
parts of the code. The idea is that this acts as a allowlist so
that the remaining cases can eventually be cleaned up and I plan
to file bugs for these cases.

LSC presubmits pass, so this shouldn't need a complicated
migration.

Change-Id: I8eee3b7b9400bf694ebc97aed7f91adbb6b52115
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/473537
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>
Reviewed-by: Pascal Perez <pascallouis@google.com>
API-Review: Pascal Perez <pascallouis@google.com>