[logs] Add E2E test for driver logging
This patch adds an E2E test for driver logging

NOTE: Because this is a host test that depends on a device target,
it cannot be just added to the dependency chain from //src:tests,
since this would create a circular dependenyc that GN will complain
loudly about.

Instead, for testing, one should add //src:host_tests to the
'host_labels' build argument, either manually or through a product
configuration. to ensure the test is properly built, with the .zbi
file it requires at runtime.

Test: fx build && ./out/default/femu_test

Change-Id: I4888b3ca852a9c015f008dd7ef7422659a4fb83f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/474621
Commit-Queue: Brian Bosak <bbosak@google.com>
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>