[kernel][arm64] Sync executable ranges prior to mapping
This moves the icache sync logic from the page fault handler into the
arch mmu code. In the process the icache sync is also moved to happen
before a mapping is installed. This removes a race condition where
another thread in the process could use the unsynced mapping.

Performing the icache sync in the mmu code, instead of the VmMapping,
corrects an existing case where sync is not done when mapping via
VmMapping::MapRange, and prevents any future chances of forgetting to
sync.

Change-Id: Icacf8f0417b94b915526b84da0146a10536a1469
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/470698
Commit-Queue: Adrian Danis <adanis@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>