[licenses] Add new zstd cases
We're cleaning up to use a verbatim github repo mirror for zstd.
The upstream project is dual-licensed and so its sources contain
both a BSD "LICENSE" file, which applies to our use, and a GPLv2
"COPYING" file, which does not apply to us.  Since we track an
exact mirror of the upstream repo, it's not an option to remove
the files we don't use from the tree altogether.

When the transition is complete, the existing entry for the
zircon/third_party/ulib/zstd directory will be removed as that
directory will no longer exist.

Bug: 68624
Change-Id: Iba17b8e56b297dd03c16e5a076a1f4ed6765b882
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/477738
Fuchsia-Auto-Submit: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Jerry Belton <jcecil@google.com>
Commit-Queue: Roland McGrath <mcgrathr@google.com>