[journal] Log failed operations
Produces logs like this (the wrapping is unfortunate but unavoidable
since so much of this log is just metadata):

```
[00008.916] 09035:09241> [00008.916191][9035][9241][/pkg/bin/minfs, journal] WARNING: [zircon/system/ulib/fs/journal/journal_writer.cc(305)] WriteOperations: Failed to write requests [BufferedOperation {type: kWrite vmo_offset: 0 dev_offset: 262146
[00008.916] 09035:09241> length: 4}]: ZX_ERR_INTERNAL. Filesystem now read-only.
```

This is useful for post-hoc analysis to understand why a journal write
may have failed; currently all we get is the status code. Journal write
failures should be unusual so the verbose output is warranted.

Also added logging in a few other error branches in the journal.

Bug: b/177663301
Tested: Logging-only enhancement
Change-Id: I5480e18ffadb9b85144e138fb74bfda43ba97ec2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/477702
Commit-Queue: James Sullivan <jfsulliv@google.com>
Reviewed-by: Chris Suter <csuter@google.com>