[netstack3] Stricter typing for IP address
In `net-types`, introduce a new `Ipv6SourceAddr` type, which
represents the source address on an IPv6 packet - such addresses
can only be unspecified or unicast, but not multicast.

Modify the IP layer API so that, when an inbound IPv6 packet is
received,
- if its source address is multicast, it is silently dropped
- otherwise, its source address is stored as `Ipv6SourceAddr`

This allows consumers of inbound IPv6 packets to rely on stricter
type bounds on the source addresses of those packets.

Also modify the `icmp` module's functions to accept `SpecifiedAddr`s
where appropriate, and push validation to the caller (the `ip`
module).

Test: fx test netstack3_core_lib_test -- \
      test_drop_non_unicast_ipv6_source

Change-Id: Ie4bcc16f8ce663a2813e44a9131c7d3516db209a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/476508
Commit-Queue: Joshua Liebow-Feeser <joshlf@google.com>
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>