[netstack] Implement error interface on pointer
Implementing interfaces on value types causes the interface to be
implemented by both the value type and the pointer type of the
implementer. This is a footgun because it makes it harder to write
correct type assertions; it requires the assertion to check for both the
pointer type and the value type.

Since interface boxing always causes the boxed object to escape, the
performance implications of implementing interfaces on pointers is nil,
and the result is better type safety.

This changes TcpIpError's error implementation to the pointer type.

Change-Id: I68a70d9095ae117380887e521bb5b88894384e09
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/477879
Fuchsia-Auto-Submit: Tamir Duberstein <tamird@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>