[brcmfmac] Refactor async_dispatcher usage in Device classes
This commit refactors the use of async_dispatcher in Device classes.  In
hardware-associated classes (PcieDevice, SdioDevice), the
async_dispatcher is provided by an owned async::Loop.  In the SimDevice,
the async_dispatcher is now refactored to use the dispatcher provided by
the wlan::simulation::Environment.

By standardizing all Device classes to provide task dispatch through an
async_dispatcher, we are able to refactor the brcmf_bus_ops interface to
remove the .set_sim_timer and .cancel_sim_timer entry points, which
allows us to simplify the Timer interface to remove its dependency on
the bus.

We are also able now to remove the Device::Init() entry point which was
previously a layering violation (required to be called by subclasses),
and to simplify the lifetime semantics of the Device and remove the
explicit Start() and Stop() entry points.

Test: fx test brcmfmac-test brcmfmac-sdio-test brcmfmac-sim-test
      Booted and connect to WLAN on Astro

Change-Id: Ieaed5efe922b5ab422fb161d6e239664cb444155
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/474824
Commit-Queue: John Sheu <sheu@google.com>
Reviewed-by: Zhiyi Chen <zhiyichen@google.com>
Reviewed-by: Sakthi Vignesh Radhakrishnan <rsakthi@google.com>