[fshost][blobfs] launch filesystems using export directory
This changes fshost to launch the filesystems using
PA_DIRECTORY_REQUEST, which makes them create an outgoing directory to
put the filesystem root and other services into. This change also
removes the temporary workaround for the diagnostics directory, it's now
available in the outgoing directory.

Today, the filesystem-mounter is in charge of pulling the appropriate
apis out of the export directory and placing them where they need to go.

Bug: 54521
Bug: 34530
Fixed: 34531
Test: fx test blobfs-integration and CQ
Change-Id: Ie817963a8526230a15edbf0e1099c6248dd5ea97
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/477610
Reviewed-by: James Sullivan <jfsulliv@google.com>
Reviewed-by: Chris Suter <csuter@google.com>
Commit-Queue: Stephen Demos <sdemos@google.com>