[brcmfmac] Fix workqueue test flakiness
The workqueue tests were susceptible to issues with spurious wakeups in
sync completions. Since a sync completion can wake up without being
signaled we also need to check some kind of condition to determine if we
should keep waiting or not.

Add a small support class that combines the sync completion with a flag
indicating if the completion was signaled or not.

BUG: 68548
Test: Ran test 100s of times with no issues
Change-Id: I1e4e897ce98b0fc82fd911abe78266a7597ffa4f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/477601
Reviewed-by: Zhiyi Chen <zhiyichen@google.com>
Commit-Queue: Bjoern Johansson <bjoernj@google.com>