[cm] Refactor routing code to be declarative
The goal of this change is to wrap the algorithm of traversing
the Component Topology into an abstraction that is flexible enough
to handle the various capability types present in CM today while
promoting code-reuse.

This is achieved by enforcing a structure on the data types that
define the shape of the topology: Use, Offer, and Expose data
structures.

With a common structure (represented by the UseDeclCommon,
OfferDeclCommon, and ExposeDeclCommon traits), the algorithm
to traverse the toplogy can be generic, with various extension
points to alter the behavior.

For example, not all capabilities can be exposed (Event, Storage),
and some capabilities cannot be sourced from a component (Event).

The declarative API uses types to build the routing algorithm at
compile time.

Finally, capabilities may need to keep track of state while they
traverse the topology. A Visitor impl is will receive
a call for each data structure visited.

NOTE: Future changes will introduce derive proc-macros to generate
impls of many of the traits introduced in this change, making for
less repetitive code and succinct implementations of each capability
type.

Change-Id: Id6a73817c5edfd7e46968345489d36f5de1eee6f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/460135
Commit-Queue: Adam Lesinski <adamlesinski@google.com>
Fuchsia-Auto-Submit: Adam Lesinski <adamlesinski@google.com>
Reviewed-by: Derek Gonyeo <dgonyeo@google.com>