[build] Remove archive targets from uefi-disk target deps
These files are not actually inputs to uefi-disk action target,
but their targets are erroneously in its deps.  Since that target
is in the default graph for x64 builds, building the archives is
in the default graph for no reason.  This change fixes that.

Change-Id: I19673b2f474c47259131afa0cd4f223f6073d9d2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/477957
Reviewed-by: Petr Hosek <phosek@google.com>
Reviewed-by: James Tucker <raggi@google.com>
Commit-Queue: Roland McGrath <mcgrathr@google.com>
Fuchsia-Auto-Submit: Roland McGrath <mcgrathr@google.com>