[fx emu][gn sdk] Fix flaky tests in femu-test.sh
The previous checks on sudo commands was flaky. The test was running
the entire script and trying to look for the last .N file to check
for a command it was expecting. Also, the start-unsecure-internet.sh
script does cat /proc/sys/net/ipv4/ip_forward and this file may not
exist or be set to 0 on some Linux machines, and would also fail the
test.

Change-Id: Iad412a9d6b2196ed431d5dc18d126861d779db15
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/478529
Commit-Queue: P.Y. Laligand <pylaligand@google.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>
Reviewed-by: P.Y. Laligand <pylaligand@google.com>