[memalloc] Allow iteration over available ranges.
Add support for iterating over the ranges currently tracked in the
allocator.

We'll shortly need this for ensuring that physical memory tracked by the
allocator has been mapped into the current address space, and a little
later on when we want to release all the unused memory in the allocator.

Change-Id: I8b8d4e6ef845dd553f3adf0cdf32750de74acc00
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/477921
Commit-Queue: David Greenaway <dgreenaway@google.com>
Fuchsia-Auto-Submit: David Greenaway <dgreenaway@google.com>
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>