[make-fuchsia-vol] Fix bugs in and usage of optimalBlockAlign
optimalBlockAlign wouldn't always return the correct values. This is
because it would incorrectly chop off the end of a partition if the
partition size was not divisible by the logical block size of the disk.
It would also incorrectly add a block to partitions that were the
logical block size of the disk.

While we're here, fix the calculation of the size of the FVM partition,
where a separate off-by-one error would result in a partition that was 1
block shorter than intended.

Change-Id: I5b2437741cd6c2587e6f47a17576a9f6163d9dbe
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/477097
Commit-Queue: Simon Shields <simonshields@google.com>
Reviewed-by: James Tucker <raggi@google.com>