[netstack/dhcp] Remove flake in tests
Commit 52021143eea968c388fdf366d843b7aa5ad371c9 introduced a change to
asynchronously deliver network packets to the network dispatcher in
tests. This removed the ability for tests to reliably synchronize with
packet delivery.

This change adds a callback to the test endpoint that is called after a
packet is successfully delivered to the network dispatcher that tests
may use to synchronize with packet delivery.

Test: fx test netstack-dhcp-gotests -- -test.count 100
MULTIPLY: netstack-dhcp-gotests: 100

Fixed: 68050
Change-Id: Ic1bed99224103e439470f457a4b5fa897aa6aa28
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/479177
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Ghanan Gowripalan <ghanan@google.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>