[scenic][fidl] Set non-nullable union in scenic_session_unittest
FIDL previously did not error during HLCPP encode when non-nullable
unions were unset, so errors were missed. Now, we are trying to
correct this behavior.

In particular, this CL fixes failing tests in
https://fuchsia-review.googlesource.com/c/fuchsia/+/477757

scenic_session_unittests encodes a non-nullable union but does
not set it. This CL sets the union and then updates the assertions
to follow from that.

Change-Id: Icadc28609fbb01f15818e5c919b8d10c59604790
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/479238
Reviewed-by: Emircan Uysaler <emircan@google.com>
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>