[fidl][llcpp] Fail subsequent calls after a client has been unbound
In Icd592da6f0e9daaed9d46d94651d6055a9f06133, the lifetime of the
channel was decoupled from the bindings on the client side (see that CL
for the background and context). That introduced a silent regression
where it was possible to make calls on an unbound client that would
seemingly succeed, except that nobody would be handling the responses
anymore.

This CL fixes that regression by still maintaining the decoupled
lifetime, but getting the channel from the binding instead of the
ChannelRefTracker when making calls.

In short, when a user |Unbind|s a client, the channel still stays alive
in the client through the ref tracker, but one cannot make any further
FIDL method calls. The client enters a "half-closed" state, ready for
eventual cleanup via its destructor.

Test: Added one in `fx test -vo fidl-walker-tests`
Change-Id: Ia9a1b11be68605e75cf93a40b2cbcbdd8cbd7d0e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/478624
Fuchsia-Auto-Submit: Yifei Teng <yifeit@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Vincent Belliard <vbelliard@google.com>