[camera] Remove references to Luis.
This bug was discovered while reproducing b/178749600.
It turns out that Luis camera controller code was still
being built if "--with //bundles:tests" was specified.

The dependency is:

//bundles:tests
  //topaz:topaz
    //topaz:tests
      //garnet/packages:garnet
        //garnet/packages/prod:all
          //garnet/packages/prod:drivers
            //garnet/packages/prod:drivers-system
              //src/camera/drivers:all
                //src/camera/drivers/controller/luis

Since Luis is no longer supported, the dependence at
//src/camera/drivers:all seems inappropriate.

There are other perspectives on how this is an issue
in the first place. See bug(s) for more details.

Test: Manually inject syntax error into luis code
      to ensure it isn't being built.
Bug: b/178973792,b/179066481
Change-Id: Idb17aab3bb9cc4bdd61114359b7190098567df80
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/479378
Commit-Queue: Ernest Hua <ernesthua@google.com>
Reviewed-by: Nic Zoghb <nzo@google.com>