[io_util, fasync] Refactor io_util and files_async libraries.
We need to bring ffx to feature parity with cs by letting ffx use cs as
a library. Since ffx is running on the host and cs uses some libraries
in io_util and files_async that are running on fuchsia, we need to add
cfg flags so that those functions running on fuchsia will only compile
for fuchsia. This CL also substitutes fuchsia_zircon::status with
fuchsia_zircon_status.

This will be necessary changes for the cs migration to work before dgonyeo@ lands the new library.

Bug: b/68555
Test: fx test files_async_tests
      fx test io_util_tests

Change-Id: If9bb7936b8f518707f53f5f997a2ebc9d300617e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/474526
Commit-Queue: Ingrid Feng <ingridlol@google.com>
Reviewed-by: Didrik Nordstrom <dnordstrom@google.com>
Reviewed-by: Stephen Demos <sdemos@google.com>
Reviewed-by: Derek Gonyeo <dgonyeo@google.com>