[kernel][pmm] Don't hold lock while iterating pages to free
The kernel sometimes frees relatively long lists of pages in a single
call to pmm_free (e.g. the page scanner during eviction).

To reduce pmm lock pressure, refactor PmmNode::FreeList to iterate
over the list to be freed prior to acquiring the lock when the pmm
checker is disabled.  When pmm checker is enabled, the list will be
iterated a second time after acquiring the lock as the pmm checker
itself is not safe for concurrent access.

This change should provide a performance improvement for the common
case where pmm checker is disabled, and a slight performance
regression when pmm checker is enabled.

Remove thread-safety annotations from AsanPoisonPage and
AsanUnpoisonPage as those functions don't actually need to be called
with the pmm lock held.  Also, make them free functions since they
don't need to access the PmmNode.

Bug: 68534
Change-Id: I8fcaa6a552012341184adcb76c25125acd658379
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/470099
Reviewed-by: Marco Vanotti <mvanotti@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Commit-Queue: Nick Maniscalco <maniscalco@google.com>