[kernel] Add support for many relocs in gen-kaslr-fixups.sh
This CL adds a workaround for ancient versions of awk that did not
support having many references to the same string in memory.

The gen-kaslr-fixups.sh script executes an awk program that, for error
reporting purposes, kept a dictionary of relocations to section
mappings. Builds that have too many relocations in the same section
(>65k) would overflow the reference counting in old versions of awk,
leading to segmentation faults.

The fix just makes sure that each entry in this map is a unique string,
by concatenating the section name with the relocation address.

Bug: 68230
Change-Id: I47a6096d54edc5d86b43d340d79d2359f1fc40bb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/478019
Commit-Queue: Marco Vanotti <mvanotti@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>