[blobfs][host] Handle null blob correctly
If the null blob was added to an initial blobfs image, its inode
erroneously had a single zero-length extent (rather than no extents).
This will eventually cause write failures if the blob was deleted on a
Fuchsia device, since the underlying block devices do not in general
support zero-length trim and/or write operations.

Bumps filesystem revision because images generated by host tooling
before this change would have zero-length extents for the null blob (so
if we find any images with oldest_revision less than the new value, they
*might* have a zero-length extent).

We automatically repair the zero-length extent on mount (when the
filesystem is mounted as writable), and update the oldest_revision to
the latest version when doing so. Note that we now have a sequence of
possible automatic upgrades that are performed on mount. Each "step"
only bumps oldest_revision by 1.

Added a check for zero-length extents to fsck.

Bug: b/177663301
Tested: Unit test; verified failed before and passed after.
Change-Id: Iea081ec1ec9b84b011991f5b9a285ba6f266320c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/478319
Commit-Queue: James Sullivan <jfsulliv@google.com>
Reviewed-by: Chris Suter <csuter@google.com>