[kernel][vm] Optionally free page tables and rename API
Extends the ArchVmAspace::FreeUnaccessed method to take an optional
parameter to control whether it just harvests accessed information, or
also reclaims the page tables. As part of this the method is renamed
to HarvestNonTerminalAccessed, since it is not always freeing.

Optional freeing allows for regularly harvesting the accessed bits
in preparation for reclamation, but only actually reclaiming when
desired, such as when actually under memory pressure.

Bug: 34565

Change-Id: I8cbe6d239345d84750644b6a2a194b937ff54a2c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/477923
Commit-Queue: Adrian Danis <adanis@google.com>
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>