[events] Remove BlockingEventSource
With this change, there is only one kind of EventSource and individual
events are either sync or async. All affected tests are updated to use
EventSource. For OpaqueTest, there is an implicit static EventStream
that listens for Resolved and is dropped when the EventSource is
dropped.

Change-Id: I08a50ea7d5deef2c1cf471c3a5dc1db072f36a19
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/474625
Commit-Queue: Fady Samuel <fsamuel@google.com>
Reviewed-by: Gary Bressler <geb@google.com>
API-Review: Gary Bressler <geb@google.com>