[session_manager] Move session_manager into core.cml
This change moves session_manager to be a child of core.cml.

This allows us to delete the component manager instance that was unique
to session manager.

The change follows the same pattern as was used for appmgr to make sure
that the system does not go down when session manager is not available.

Fixed: 64080

Change-Id: Iafd9bb78d7adc50f29b85c69a0842b45e3efad3c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/471639
Commit-Queue: Kevin Lindkvist <lindkvist@google.com>
API-Review: Kevin Lindkvist <lindkvist@google.com>
Reviewed-by: Rich Kadel <richkadel@google.com>
Reviewed-by: Gary Bressler <geb@google.com>
Reviewed-by: Chase Latta <chaselatta@google.com>