[sdk] Temporarily disable e2e tests
These tests run on an infra builder that does not use the builder-tester
split, which is preventing infra from optimizing CQ->GI time by switching
to different VMs. They currently hang on the new VMs. We have no way of
testing fixes on the new VMs until infra rolls them out, so we we'll
disable the test for now and attempt to reland when the new VMs are
configured.

Bug: 68296
Change-Id: I5f4bb25006b35d0d537c84123a1451b403aaec08
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/480597
Reviewed-by: Nathan Mulcahey <nmulcahey@google.com>
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Commit-Queue: Kendal Harland <kjharland@google.com>