[fvm] Fix partition size limiting.
The device path used by fshost for FVM was incorrect so size limits
were not being applied.

Add more logging to FVM failures.

Adds additional logging for failed partition lookups, especially around
partition size limits. There was a report of a failure and it's not
clear why FVM is issuing the error.

Separate out the FshostIntegrationTest fixture implementation into a .cc
file and move the waiting code from the recovery-test to there so it can
be shared with the new test. Rename the test fixture file to match the
class name.

Reorganize the declarations of the fshost config constants so they are
easier to follow and the documentation can be longer.

Bug: 64055

Change-Id: I847d95c50d648551212bdd348875ab9f2317160f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/475442
Commit-Queue: Brett Wilson <brettw@google.com>
Reviewed-by: Chris Suter <csuter@google.com>