[bt][a2dp] Detach tests from AudioConsumerFactory
Always require a SessionAudioConsumerFactory when building a Player.

Encoding tests previously tried to connect to play audio, even though
it is not available in the test environment. Provide the factory proxy
when it is needed through the SinkTaskBuilder.

Bug: 43584
Test: updated tests in bt-a2dp-encoding-tests
      manually observe test framework not log about a failed connect

Change-Id: I01bb809ada318fd701b8583520482b6e53c91e32
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/480217
Commit-Queue: Marie Janssen 💖 <jamuraa@google.com>
Fuchsia-Auto-Submit: Marie Janssen 💖 <jamuraa@google.com>
Reviewed-by: Jeff Belgum <belgum@google.com>