[wlan][policy] Connect to WPA3 with WPA config
This change allows network configs saved as WPA to be used to connect to
observed WPA3 networks when auto connecting.

This change does not allow networks saved as WPA3 to be used any
differently.

Security type is not used as a determining factor in
network selection, so if we see two APs with the same SSID and WPA and
WPA3 security, both will match with a config saved as WPA and the actual
security type will not be used to select a network.

Bug: 472909
Test: Added a unit test and modified existing tests:
    fx test wlancfg-tests
  Manually tested that a network saved through the donut tool as
  WPA will trigger a connect to a WPA3 network.
Multiply: wlancfg-tests

Change-Id: I6c663e18a2e60a9b90a609238f46af06f8445239
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/480009
Reviewed-by: Kevin Sakuma <sakuma@google.com>
Reviewed-by: Marc Khouri <mnck@google.com>
Commit-Queue: Naomi Mccracken <nmccracken@google.com>