[fx] Create a root .fx/ dir for caching artifacts between runs
For now, this is only used for caching the new fx-set Go executable.
Eventually it's probably a good place to centralize the information
that's currently in one-off files like .fx-build-dir.

It's already in .gitignore so it shouldn't be noticeable to any users.

Bug: 68465
Change-Id: Ieeacede45360b4cea3db199c5bdfdf4c88260719
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/480048
Fuchsia-Auto-Submit: Oliver Newman <olivernewman@google.com>
Reviewed-by: Renato Mangini Dias <mangini@google.com>
Reviewed-by: Gary Boone <gboone@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>