[brcmfmac] Add RecoveryTrigger for driver recovery
This change introduced RecoveryTrigger class as the central controller
of start points of the recovery process, and TriggerCondition class to
keep track of the instaces of different kinds of errors, there are only
two conditions for now. Also this change moved the recovery worker to
core layer.

Bug: 59364
Test: 1. New unittests in brcmfmac/recovery/test/recovery_trigger_test.cc
      2. fxr/434158

Change-Id: I784812091bd9d928331b07d76758c0ed0beb0ca1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/466254
Commit-Queue: Zhiyi Chen <zhiyichen@google.com>
Reviewed-by: Sakthi Vignesh Radhakrishnan <rsakthi@google.com>