[pkg-cache] Implement fuchsia.pkg.PackageCache/Get
This change replaces the existing stub handler for the
fuchsia.pkg.PackageCache/Get API with a functional implementation.

While it is currently implemented using pkgfs, a future change can
replace its implementation with one based on blobfs without needing to
change the FIDL contract.

Most of the logic in pkg-cache to serve this API (and the NeededBlobs
protocol) was in prior CLs, though this is the first one to change the
public API of pkg-cache. So, this change also starts on pkg-cache
integration tests that cache packages through pkg-cache instead of
pkgfs.

Fixed: 64624
Change-Id: I20142edd4a69c80745905848dc1bf31a60a46825
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/475677
Reviewed-by: Erick Tryzelaar <etryzelaar@google.com>
Reviewed-by: Sen Jiang <senj@google.com>
Commit-Queue: Kevin Wells <kevinwells@google.com>