[cml][runner] Remove `from` property of RunnerDecl
The `from` property of the RunnerDecl was initially added to
add another round of routing, so that a protocol could be routed
to a `runner` capability, which is then routed to an environment.

This indirection is unnecessary and routing can always be performed
from the source component in the form of a `runner` capability.

Bug: 52195
Change-Id: Ice463cbd7fd660224e52bcc599a8895706c5919e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/479197
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Reviewed-by: Gary Bressler <geb@google.com>
Commit-Queue: Adam Lesinski <adamlesinski@google.com>
Fuchsia-Auto-Submit: Adam Lesinski <adamlesinski@google.com>
API-Review: Gary Bressler <geb@google.com>