[sshutil] Don't hold lock while reconnecting
Previously, `Client.Reconnect()` would hold the client lock while
reconnecting. This would cause all the other `Client` methods to block
until we stopped attempting to reconnect (which might be never).  Most
significantly would be `Client.Close()`, which typically gets called
with a `defer client.Close()` while unwinding because of an error.

This fixes this issue by not holding the lock while trying to connect to
the target device. This has the unfortunate side effect that it's
possible to have multiple reconnection attempts in flight, so this patch
detects this case, and just closes the excess connections.

Change-Id: Ic82e7998ef0c088f03dc43a94130afd7a494fed5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/480148
Fuchsia-Auto-Submit: Erick Tryzelaar <etryzelaar@google.com>
Commit-Queue: Erick Tryzelaar <etryzelaar@google.com>
Reviewed-by: Oliver Newman <olivernewman@google.com>
Reviewed-by: Kevin Wells <kevinwells@google.com>