[zbi] Update ZBI spec for modern booting design
This changes the wording to remove the "complete ZBI" term
and instead refer only to a "bootable ZBI".  The definition
is simplified to require only the kernel item, with the other
items required left to the particular kernel's needs.

The zbi tool is updated to replace --complete with --bootable
and its check simplified to match the new definition.

Also add the new type ZBI_TYPE_STORAGE_KERNEL, which will be used
in the new "physboot" booting design for Zircon.

Modernize the formatting of the header file.

Bug: 57245, 68585
Change-Id: I4497cf2989df902382d6f7546b67995fe547449c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/480364
Commit-Queue: Roland McGrath <mcgrathr@google.com>
Reviewed-by: David Greenaway <dgreenaway@google.com>
Fuchsia-Auto-Submit: Roland McGrath <mcgrathr@google.com>