[fidl][llcpp] Move LLCPP wire types into wire::
This moves type declarations for wire types into the wire::
sub-namespace but introduces using aliases for them in their original
location. This allow the first phase of this change with (virtually)
no changes to user code. User code can then be migrated to refer to the
new location smoothly before the alias is removed.

Bug: 65964

Change-Id: I63310ff0c8644911139890aa2c0ce359aaec99f3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/475721
Commit-Queue: Ian McKellar <ianloic@google.com>
Reviewed-by: Vincent Belliard <vbelliard@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>