[components] Modify directory offer to use "as"
It looks like there was a typo in the example on how to define offers. In the "directory" example, it
lists "directory" twice, but it appears that this section was trying to explain how offers could be
renamed with "as", so I switched to using that.

Change-Id: Ibd0a3cefffa1d26a072bb71bd3d9ccf8b5aa2727
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/462495
Reviewed-by: Adam Lesinski <adamlesinski@google.com>
Commit-Queue: Adam Lesinski <adamlesinski@google.com>