[fidl][fidlconv] Add working proof of concept for ConvertingTreeVisitor.
This change serves as a minimal implementation of the chosen design for
the fidlconv converter.  It contains converters for (most) types, struct
declarations, struct members, and name-type ordering (ie, "uint8 FOO" ->
"FOO uint8").  More converters (and tests for them!) will be added in
future changes.

The basic idea of the design is to encapsulate every type of transform
specified by FTP-050 into its own "Conversion" class.  Some of these
transforms can be nested within other transforms (example: "alias foo =
array<handle?>:5" contains a transform of the "handle" type definition
within the transform of the "array" type definition).  Because of this
property, the ConvertingTreeVisitor stores Conversions on a stack,
taking the output of each previous transform and adding it as a "child"
to its parent.  When the stack is exhausted, the final result is written
to an output string containing the text of the fully converted file.

Test: fidldev test fidlc
Bug: 68667
Change-Id: I5e7573c93df18ec8802f8f09a54e6fd0ddd2076c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/474707
Reviewed-by: Felix Zhu <fcz@google.com>
Commit-Queue: Alex Zaslavsky <azaslavsky@google.com>