[audio_core] Remove LinearSampler
This CL removes LinearSamplerImpl from audio_core. If an inexpensive
non-Point sampler is needed in the future, the windowed-sinc sampler
can be configured with a narrower width; in fact the LinearFilter
subclass already represents most of the required work.

The removal saves about 152 KiB in binary size, without any impact
to the product.  LinearSamplerImpl cannot be directly selected by
API, and internally only unittests explicitly select it. Because
all references from these tests are being removed or changed, it is
safe to remove the LinearSampler sources at this time.

Test: fx test -o audio_mixer_unittests audio_fidelity_tests

Change-Id: Ida05fd868b14584d1cb7cc7b4b987e1184ac60e7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/478486
Commit-Queue: Martin Puryear <mpuryear@google.com>
Reviewed-by: Tom Bergan <tombergan@google.com>